+++ cmd/datawatch/main.go
@@ -84,7 +84,7 @@ import (
 )
 
 // Version is set at build time via -ldflags.
-var Version = "4.0.2"
+var Version = "4.0.3"
 
 var (
 	cfgPath    string